Invalidate cache instead of updating it
Asynchronously provided changes given to handle_changes()
can't be reconciled with the local cache because there is
no way to tell which data is more recent. This results
in files temporarily reverting to an older state. This causes
issues with programs that assume no other program is editing
their files. Instead of trying to keep cache in sync with remote
changes, just invalidate cache. This forces us to retrieve new
metadata when edits are made.