Don't create cache counter when in read-only mode
_get_stat_num() creates the cache counter if it didn't exist,
signalling an intent to update that cache entry. This was being done
when the DB connection was in a read-lock which would cause it to fail
intermittently.  Now, we optimistically get the cached metadata
without getting the cache counter and if it turns out we have to
write, then try again with a read-write lock.