Retry stat if file was closed during retrieval
If an open file was written/closed while we retrieved its stat, that
stat rev could be out of date. This would allow the possibility for a
file that was only edited locally to revert one version back if we
re-opened it immediately after we closed it.

This change forces a re-retrieval of a file's stat if any file that
was written was closed while we were retrieving that stat. This
prevents the possibility of sequential local edits from being lost.