Release GIL before calling sqlite3_close()
Current versions of pysqlite erroneously call sqlite3_close() without
releasing the GIL. This can cause a deadlock if sqlite3_close must
acquire a lock internal to sqlite3 that another thread may currently
hold.  The thread currently holding the sqlite3-internal may never be
able to release the lock if it must first acquire the GIL.

This workaround should continue to work when this is fixed upstream.

Fixes #45